Ispravio za poslovnice delove, pogledaj što ne prikazuje kvartove iako ih povuče, iskucaj za ugovore šta i koliko stigneš. Jebavam se i ja za ova Lica, ne znam zašto neće. Promenio da su JMBGPIB stringovi da mogu i 0 da počnu

+++ AgencijaNekretnine/Entiteti/Zaposleni.cs
@@ -8,7 +8,7 @@ namespace AgencijaNekretnine.Entiteti
 {
    public class Zaposleni
     {
-        public virtual int JMBG { get; set; }
+        public virtual long JMBG { get; set; }
         public virtual String Ime { get; set; }
         public virtual String Prezime { get; set; }
         public virtual DateTime DatZaposlenja { get; set; }
